backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
tensorflow==2.15.0
pillow==10.2.0
python-multipart==0.0.9
google-generativeai==0.4.1
python-dotenv==1.0.1
pydantic==2.6.4